requirements.txt
Flask>=2.0
Flask-SocketIO>=5.0
python-dotenv>=0.19
openai>=1.0.0